indexmap_nostd

Module map

An ordered map based on a B-Tree that keeps insertion order of elements.

Structs

IndexMap

A b-tree map where the iteration order of the key-value pairs is independent of the ordering of the keys.

IntoIter

An owning iterator over the entries of a IndexMap.

Iter

An iterator over the entries of an IndexMap.

IterMut

A mutable iterator over the entries of an IndexMap.

OccupiedEntry

A view into an occupied entry in a IndexMap. It is part of the Entry enum.

VacantEntry

A view into a vacant entry in an IndexMap. It is part of the Entry enum.

Values

An iterator over the values of an IndexMap.

ValuesMut

An iterator over the values of an IndexMap.

Enums

Entry

A view into a single entry in a map, which may either be vacant or occupied.