Struct indexmap_nostd::set::IndexSet

pub struct IndexSet<T> { /* private fields */ }

A b-tree set where the iteration order of the values is independent of the ordering of the values.

The interface is closely compatible with the indexmap crate and a subset of the features that is relevant for the wasmparser-nostd crate.

Differences to original IndexSet

Since the goal of this crate was to maintain a simple no_std compatible fork of the indexmap crate there are some downsides and differences.

Some operations such as IndexSet::insert now require K: Clone.

It is to be expected that this fork performs worse than the original indexmap crate implementation.

The implementation is based on BTreeMap internally instead of HashMap which has the effect that methods no longer require K: Hash but K: Ord instead.

Implementations

impl<T> IndexSet<T>

pub fn new() -> Self

Makes a new, empty IndexSet.

Does not allocate anything on its own.

pub fn with_capacity(capacity: usize) -> Self

Constructs a new, empty IndexSet with at least the specified capacity.

Does not allocate if capacity is zero.

pub fn reserve(&mut self, additional: usize)

Reserve capacity for at least additional more values.

pub fn len(&self) -> usize

Returns the number of elements in the set.

pub fn is_empty(&self) -> bool

Returns true if the set contains no elements.

pub fn is_disjoint(&self, other: &Self) -> bool
where T: Ord,

Returns true if self has no elements in common with other. This is equivalent to checking for an empty intersection.

pub fn is_subset(&self, other: &Self) -> bool
where T: Ord,

Returns true if the set is a subset of another, i.e., other contains at least all the elements in self.

pub fn is_superset(&self, other: &Self) -> bool
where T: Ord,

Returns true if the set is a superset of another, i.e., self contains at least all the elements in other.

pub fn contains<Q>(&self, key: &Q) -> bool
where T: Borrow<Q> + Ord, Q: Ord + ?Sized,

Returns true if the set contains an element equal to the value.

The value may be any borrowed form of the setโs element type, but the ordering on the borrowed form must match the ordering on the element type.

pub fn get<Q>(&self, value: &Q) -> Option<&T>
where T: Borrow<Q> + Ord, Q: Ord + ?Sized,

Returns a reference to the element in the set, if any, that is equal to the value.

The value may be any borrowed form of the setโs element type, but the ordering on the borrowed form must match the ordering on the element type.

pub fn get_full<Q>(&self, key: &Q) -> Option<(usize, &T)>
where T: Borrow<Q> + Ord, Q: Ord + ?Sized,

Returns the index-value pair corresponding to the supplied value.

The supplied key may be any borrowed form of the mapโs key type, but the ordering on the borrowed form must match the ordering on the key type.

pub fn get_index_of<Q>(&self, key: &Q) -> Option<usize>
where T: Borrow<Q> + Ord, Q: Ord + ?Sized,

Returns the unique index corresponding to the supplied value.

The supplied key may be any borrowed form of the mapโs key type, but the ordering on the borrowed form must match the ordering on the key type.

pub fn get_index(&self, index: usize) -> Option<&T>

Returns a shared reference to the value at the given index.

pub fn insert(&mut self, value: T) -> bool
where T: Ord + Clone,

Adds a value to the set.

Returns whether the value was newly inserted. That is:

If the set did not previously contain an equal value, true is returned.

If the set already contained an equal value, false is returned, and the entry is not updated.

pub fn insert_full(&mut self, value: T) -> (usize, bool)
where T: Ord + Clone,

Adds a value to the set.

Returns the unique index to the value as well as a bool flag telling whether the value was newly inserted. That is:

If the set did not previously contain an equal value, true is returned.

If the set already contained an equal value, false is returned, and the entry is not updated.

pub fn iter(&self) -> Iter<'_, T>

Gets an iterator that visits the elements in the IndexSet in the order in which they have been inserted into the set unless there have been removals.

pub fn clear(&mut self)

Clears the set, removing all elements.